refactors and adds types to custom hooks
